Update network responses to support 64-bit Job IDs

Closes #28151

See merge request gitlab-org/gitlab-runner!3346

+++ executors/custom/custom_test.go
@@ -117,9 +117,9 @@ func prepareExecutor(t *testing.T, tt executorTestCase) (*executor, common.Execu
 	return e, options, out
 }
 
-var currentJobID = 0
+var currentJobID = int64(0)
 
-func jobID() int {
+func jobID() int64 {
 	i := currentJobID
 	currentJobID++
